requirements.txt
torch
scikit-learn
scipy
numpy
joblib
threadpoolctl
sympy
networkx
transformers
sentence-transformers
huggingface-hub
regex
pandas
python-dateutil
pytz
PyYAML
packaging
fsspec
lxml
matplotlib
seaborn
requests
httpx
httpcore
duckduckgo-search
ddgs
urllib3
idna
certifi
python-dotenv
tqdm
filelock
typing_extensions
Jinja2
newspaper3k